Changes for the basin mesh generation.

+++ src/data/greenland/download_greenland_data.py
@@ -5,10 +5,11 @@ import zipfile
 import tarfile
 import inspect
 
-def download_file(url, direc, folder, extract=False):
+def download_file(url, direc, folder, extract=False, name = None):
   """
   download a file with url <url> into directory <direc>/<folder>.  If <extract>
   is True, extract the .zip file into the directory and delete the .zip file.
+  If name is specified, the file will be changed to the string passed as name.
   """
   # make the directory if needed :
   direc = direc + '/' + folder + '/'
@@ -18,34 +19,46 @@ def download_file(url, direc, folder, extract=False):
 
   # url file info :
   fn   = url.split('/')[-1]
-  u    = urllib2.urlopen(url)
-  f    = open(direc + fn, 'wb')
-  meta = u.info()
-  fs   = int(meta.getheaders("Content-Length")[0])
-  
-  s    = "Downloading: %s Bytes: %s" % (fn, fs)
-  print s
-  
-  fs_dl  = 0
-  blk_sz = 8192
-  
-  # download the file and print status :
-  while True:
-    buffer = u.read(blk_sz)
-    if not buffer:
-      break
-  
-    fs_dl += len(buffer)
-    f.write(buffer)
-    status = r"%10d  [%3.2f%%]" % (fs_dl, fs_dl * 100. / fs)
-    status = status + chr(8)*(len(status)+1)
-    sys.stdout.write(status)
-    sys.stdout.flush()
-  
-  f.close()
-  
+
+  # Only download files that don't exist. 
+  # This will allow a restart of a terminated download
+  if name:
+      fn = name
+
+  EXISTS = os.path.isfile(direc+fn)
+    
+  if not EXISTS:
+      u    = urllib2.urlopen(url)
+      f    = open(direc + fn, 'wb')
+      meta = u.info()
+      fs   = int(meta.getheaders("Content-Length")[0])
+      
+      s    = "Downloading: %s Bytes: %s" % (fn, fs)
+      print s
+      
+      fs_dl  = 0
+      blk_sz = 8192
+
+      
+      # download the file and print status :
+      while True:
+        buffer = u.read(blk_sz)
+        if not buffer:
+          break
+      
+        fs_dl += len(buffer)
+        f.write(buffer)
+        status = r"%10d  [%3.2f%%]" % (fs_dl, fs_dl * 100. / fs)
+        status = status + chr(8)*(len(status)+1)
+        sys.stdout.write(status)
+        sys.stdout.flush()
+      
+      f.close()
+  else:
+    print "WARNING: "+fn+" already downloaded. \nDelete file if it is was partial download"
+
   # extract the zip/tar.gz file if necessary :
-  if extract:
+  if extract and not EXISTS:
     ty = fn.split('.')[-1]
     if ty == 'zip':
       cf = zipfile.ZipFile(direc + fn)
@@ -54,6 +67,9 @@ def download_file(url, direc, folder, extract=False):
     cf.extractall(direc)
     os.remove(direc + fn)
 
+  if name and not EXISTS:
+    os.rename(direc+fn,direc+name)
+  
 def convert_measures_projection(direc, var):
   """
   convert the measures .tif files to _new.tif files with the projection we 
@@ -107,6 +123,13 @@ download_file(q_geo, home, fldr)
 searise = 'http://websrv.cs.umt.edu/isis/images/e/e9/Greenland_5km_dev1.2.nc'
 fldr    = 'searise'
 download_file(searise, home, fldr)
+
+# NASA basins dataset for Greenland:
+basins_shape  = 'https://umt.box.com/shared/static/s18gkw5k7ldozg0l9kfreh3hsxk0flpq.txt'
+basins_image  = 'https://umt.box.com/shared/static/pgo62nopbwid6hkuc5suyvggjozvywqv.png'
+fldr    = 'basins'
+download_file(basins_shape, home, fldr ,name = 'GrnDrainageSystems_Ekholm.txt')
+download_file(basins_image, home, fldr ,name = 'Grn_Drainage_Systems.png')
 
 # smooth target matlab matrix :
 smooth  = 'https://dl.dropboxusercontent.com/s/e8r0x37mil03hvu/' +\
